Move photo view out of UI, add margin to bottom bar selected photo
