+++ src/components/PhotoViewModal/PhotoViewModal.module.css
@@ -266,6 +266,8 @@
 .selectedStripPhoto {
   aspect-ratio: 1 / 1;
   opacity: 1;
+  margin-left: var(--size-space-100);
+  margin-right: var(--size-space-100);
 }
 
 .stripSkeleton {
