+++ src/components/PhotoViewModal/PhotoViewModal.tsx
@@ -3,10 +3,10 @@ import { Dialog } from "radix-ui";
 import styles from "./PhotoViewModal.module.css";
 import { AnimatePresence, motion } from "framer-motion";
 import { ReactNode, useEffect, useRef, useState } from "react";
-import { IconButton } from "../input/IconButton";
+import { IconButton } from "@/components/ui/input/IconButton";
 import { Check, Download, Info, Link, PanelLeft, X } from "lucide-react";
 import { useFlatGoodPhotos } from "@/hooks/photo";
-import { PhotoDto } from "../../../../generated";
+import { PhotoDto } from "../../../generated";
 import { useRouter } from "@tanstack/react-router";
 import { useCopyToClipboard } from "@/hooks/clipboard";
 import { usePhotoDownload } from "@/hooks/photoDownload";
